refactor ui e2e trace fixtures
